requirements.txt
absl-py==0.12.0
astunparse==1.6.3
attrs==21.2.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
clang==5.0
cloudpickle==2.0.0
colorama==0.4.4
cycler==0.10.0
Cython==0.29.24
decorator==5.1.0
dill==0.3.4
dm-tree==0.1.6
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gin-config==0.4.0
google-api-core==2.0.1
google-api-python-client==2.21.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.40.0
h5py==3.1.0
httplib2==0.19.1
idna==3.2
imageio==2.9.0
importlib-metadata==4.8.1
importlib-resources==5.2.2
joblib==1.0.1
kaggle==1.5.12
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
meta-dataset==0.2.0
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.1
opencv-python-headless==4.5.3.56
opt-einsum==3.3.0
pandas==1.3.3
Pillow==8.3.2
portalocker==2.3.2
promise==2.3
protobuf==3.18.0
psutil==5.8.0
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacrebleu==2.0.0
scikit-learn==0.24.2
scipy==1.7.1
sentencepiece==0.1.96
seqeval==1.2.2
six==1.15.0
sklearn==0.0
tabulate==0.8.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-addons==0.14.0
tensorflow-datasets==4.4.0
tensorflow-estimator==2.6.0
tensorflow-gpu==2.6.0
tensorflow-hub==0.12.0
tensorflow-metadata==1.2.0
tensorflow-model-optimization==0.6.0
tensorflow-probability==0.7.0
tensorflow-text==2.6.0
termcolor==1.1.0
text-unidecode==1.3
tf-models-official==2.6.0
tf-slim==1.1.0
threadpoolctl==2.2.0
torch==1.9.0
torchvision==0.10.0
tqdm==4.62.2
typeguard==2.12.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
